+++ R/list-modify.R
@@ -13,7 +13,8 @@
 #' method for lists.
 #'
 #' @param .x List to modify.
-#' @param ... New values of a list. Use `NULL` to remove values.
+#' @param ... New values of a list. Use `zap()` to remove values.
+#'
 #'   These values should be either all named or all unnamed. When
 #'   inputs are all named, they are matched to `.x` by name. When they
 #'   are all unnamed, they are matched positionally.
@@ -31,8 +32,9 @@
 #' # Replace values
 #' str(list_modify(x, z = 5))
 #' str(list_modify(x, z = list(a = 1:5)))
+#'
 #' # Remove values
-#' str(list_modify(x, z = NULL))
+#' str(list_modify(x, z = zap()))
 #'
 #' # Combine values
 #' str(list_merge(x, x = 11, z = list(a = 2:5, c = 3)))
@@ -40,7 +42,7 @@
 #'
 #' # All these functions support tidy dots features. Use !!! to splice
 #' # a list of arguments:
-#' l <- list(new = 1, y = NULL, z = 5)
+#' l <- list(new = 1, y = zap(), z = 5)
 #' str(list_modify(x, !!!l))
 list_modify <- function(.x, ...) {
   list_recurse(.x, list2(...), function(x, y) y)
@@ -72,7 +74,7 @@ list_recurse <- function(x, y, base_case) {
       if (i <= length(x) && is_list(x[[i]]) && is_list(y[[i]])) {
         x[[i]] <- list_recurse(x[[i]], y[[i]], base_case)
       } else {
-        x[[i]] <- base_case(x[[i]], y[[i]])
+        x[[i]] <- maybe_zap(base_case(x[[i]], y[[i]]))
       }
     }
   } else {
@@ -81,12 +83,37 @@ list_recurse <- function(x, y, base_case) {
       if (has_name(x, nm) && is_list(x[[nm]]) && is_list(y[[i]])) {
         x[[nm]] <- list_recurse(x[[nm]], y[[i]], base_case)
       } else {
-        x[[nm]] <- base_case(x[[nm]], y[[i]])
+        x[[nm]] <- maybe_zap(base_case(x[[nm]], y[[i]]))
       }
     }
   }
 
   x
+}
+
+maybe_zap <- function(x) {
+  if (is_zap(x)) {
+    return(NULL)
+  }
+  if (!is_null(x)) {
+    return(x)
+  }
+
+  signal_soft_deprecated(paste_line(
+    "Removing elements with `NULL` is soft-deprecated as of purrr 0.3.0.",
+    "Please use `zap()` instead of `NULL`",
+    "",
+    "  # Before:",
+    "  list_modify(x, foo = NULL)",
+    "",
+    "  # After:",
+    "  list_modify(x, foo = zap())"
+  ))
+
+  # Allow removing with `NULL` for now. In purrr 0.5.0, this
+  # functionality will be defunct and we'll allow setting elements to
+  # `NULL`.
+  NULL
 }
 
 #' @rdname list_modify